+++ src/shared/ui/dropdown/styles.module.css
@@ -55,7 +55,22 @@
   justify-content: space-between;
   align-items: center;
 }
-
+.itemSelected {
+  color: var(--colors-interface-additional);
+  height: 48px;
+  line-height: 46px;
+  cursor: pointer;
+  transition: all 0.2s;
+  padding: 0 10px;
+  box-sizing: border-box;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: solid 1px var(--colors-interface-additional);
+}
+.itemSelected:hover {
+  background: #f4f4f4;
+}
 .item:hover {
   background: #f4f4f4;
 }
